dedicated files for mnemonic and pubkey viewers
