+++ tests/pages/test_tiny_seed.py
@@ -3,7 +3,7 @@ from .test_home import tdata, create_ctx
 
 
 def test_export_mnemonic_tiny_seed_menu(mocker, m5stickv, tdata):
-    from krux.pages.home import Home
+    from krux.pages.mnemonic_view import MnemonicsView
     from krux.wallet import Wallet
     from krux.input import BUTTON_ENTER, BUTTON_PAGE
 
@@ -28,10 +28,10 @@ def test_export_mnemonic_tiny_seed_menu(mocker, m5stickv, tdata):
         ],
     ]
     ctx = create_ctx(mocker, case[2], case[0], case[1])
-    home = Home(ctx)
-    mocker.spy(home, "tiny_seed")
-    home.mnemonic()
-    home.tiny_seed.assert_called_once()
+    mnemonics = MnemonicsView(ctx)
+    mocker.spy(mnemonics, "tiny_seed")
+    mnemonics.mnemonic()
+    mnemonics.tiny_seed.assert_called_once()
     assert ctx.input.wait_for_button.call_count == len(case[2])
